% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.to.excel.R
\name{copy.to.excel}
\alias{copy.to.excel}
\title{Copy a data frame to Excel}
\usage{
copy.to.excel(x, row.names = FALSE, col.names = TRUE, quote = FALSE)
}
\arguments{
\item{row.names}{whether your output table should include row names (defaults to FALSE)}

\item{col.names}{whether your output table should include column names (defaults to TRUE)}

\item{quote}{whether your output should include quotation marks as field delimiters (defaults to FALSE)}
}
\description{
For a data frame in your R environment, this function will copy it to the clipboard in a format that can be pasted into Excel.
In the current implementation, this function only works on a Mac.
}
\author{
Michael C. Saul
}

