% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeo.overlap.test.R
\name{hypergeo.overlap.test}
\alias{hypergeo.overlap.test}
\title{Hypergeometric test of overlap function}
\usage{
hypergeo.overlap.test(list1, list2, background1, background2,
  representation = "over")
}
\arguments{
\item{list1}{the first hypergeometric test list}

\item{list2}{the second hypergeometric test list}

\item{background1}{the background for the first list}

\item{background2}{the background for the second list}

\item{representation}{Test for overrepresentation or underrepresentation? Defaults to "over".}
}
\description{
For four character vectors (list1, list2, background1, background2), this function computes a hypergeometric p-value of the overlap.
It can test for either overrepresentation or underrepresention.
}
\examples{
x = c("one", "two", "three", "four", "five")
y = c("two", "three", "four", "five", "six")
x.bg = c("one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten", "eleven")
y.bg = c("one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten", "twelve")
# should return p = 0.003968
hypergeo.overlap.test(x, y, x.bg, y.bg, representation = "over")
}

