% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Overlap function}
\usage{
overlap(vector1, vector2)
}
\arguments{
\item{vector1}{the first character vector}

\item{vector2}{the second character vector}
}
\description{
This function finds the overlap between two character vectors
}
\examples{
x = c("one", "two", "three", "four", "five")
y = c("two", "four", "six", "eight", "ten")
# should return c("two", "four")
overlap(x,y)
}
\author{
Michael C. Saul
}

