% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.from.excel.R
\name{paste.from.excel}
\alias{paste.from.excel}
\title{Paste data frame from Excel}
\usage{
paste.from.excel(header = TRUE, stringsAsFactors = FALSE)
}
\arguments{
\item{header}{whether your table has a header (defaults to TRUE)}

\item{stringsAsFactors}{whether character vectors should be converted to factors (defaults to FALSE)}

\item{representation}{test for overrepresentation or underrepresentation (defaults to "over")}
}
\description{
For a data frame copied from Excel, this function will paste it into the R environment.
In the current implementation, this function only works on a Mac.
}
\author{
Michael C. Saul
}

