% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSET.overlap.test.R
\name{MSET.overlap.test}
\alias{MSET.overlap.test}
\title{Modular single-set enrichment tool (MSET): randomization-based test for list over- or under-representation}
\usage{
MSET.overlap.test(list1, list2, background1, background2, B = 30000,
  representation = "over")
}
\arguments{
\item{list1}{the first test list}

\item{list2}{the second test list}

\item{background1}{the background for the first list}

\item{background2}{the background for the second list}

\item{representation}{Test for overrepresentation or underrepresentation? Defaults to "over".}
}
\description{
MSET was developed to compare gene lists.
For four character vectors (list1, list2, background1, background2), this function computes a randomization-based p-value of the overlap.
It can test for either overrepresentation or underrepresention.
MSET is based on work from Eisinger et al., 2013, "Development of a versatile enrichment analysis tool reveals associations between the maternal brain and mental health disorders, including autism." BMC Neuroscience.
}
\examples{
x = c("one", "two", "three", "four", "five")
y = c("two", "four", "six", "eight", "ten")
x.bg = c("one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten", "eleven")
y.bg = c("one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten", "twelve")
MSET.overlap.test(x, y, x.bg, y.bg)
}

